pyproject.toml
[tool.poetry]
name = "django-api-example"
version = "0.1.0"
description = ""
authors = ["Adi Utama <neogatotkaca@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.9"
django = "^3.1.1"
djangorestframework = "^3.11.1"
djangorestframework-simplejwt = "^4.4.0"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
